Fix component update issue by including NgZone
I'm not 100% sure what NgZone is, but angular only knows the state of a component
changed when it's happening inside of the component or NgZone. So I added an NgZone
object to the constructor and when I attatch the callback to gapi.load I have the videos state update
inside that NgZone object.